% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add.gaps.R
\docType{methods}
\name{add.gaps}
\alias{add.gaps}
\title{Add gap-only sequences for missing data}
\usage{
add.gaps(x, ...)
}
\arguments{
\item{x}{a \linkS4class{multidna} or \linkS4class{multiphyDat} object.}

\item{...}{further arguments passed to other methods (currently not used).}
}
\description{
In \linkS4class{multidna} and \linkS4class{multiphyDat}, some individuals may not be sequenced for all genes.
The generic function \code{add.gap} has method for both objects; it identifies the missing sequences, and adds gap-only sequences to the alignments wherever needed.
}

