% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{package}
\name{apex-package}
\alias{apex}
\alias{apex-package}
\title{apex: Phylogenetic Methods for Multiple Gene Data}
\description{
Toolkit for the analysis of multiple gene data (Jombart et al. 2017) \doi{10.1111/1755-0998.12567}. 'apex' implements the new S4 classes 'multidna', 'multiphyDat' and associated methods to handle aligned DNA sequences from multiple genes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thibautjombart/apex}
  \item Report bugs at \url{https://github.com/thibautjombart/apex/issues}
}

}
\author{
\strong{Maintainer}: Klaus Schliep \email{klaus.schliep@gmail.com} (\href{https://orcid.org/0000-0003-2941-0161}{ORCID})

Authors:
\itemize{
  \item Thibaut Jombart \email{t.jombart@imperial.ac.uk}
  \item Zhian Namir Kamvar \email{kamvarz@science.oregonstate.edu}
  \item Eric Archer \email{eric.archer@noaa.gov}
  \item Rebecca Harris \email{rbharris@uw.edu}
}

}
\keyword{internal}
