% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exports.R
\name{multidna2genind}
\alias{multidna2genind}
\title{Convert multidna into genind}
\usage{
multidna2genind(x, genes = TRUE)
}
\arguments{
\item{x}{a \linkS4class{multidna} object.}

\item{genes}{an optional vector indicating the genes to retain for the concatenation; any way to subset the list in x@dna is acceptable; by default, all genes are used.}
}
\value{
a \linkS4class{genind} object
}
\description{
This function concatenates separate DNA alignments, and then extracts SNPs of the resulting alignment into a \linkS4class{genind} object.
}
\examples{
## simple conversion with nicely ordered output
data(woodmouse)
genes <- list(gene1=woodmouse[,1:500], gene2=woodmouse[,501:965])
x <- new("multidna", genes)
x
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}
\seealso{
\itemize{
\item concatenate
\item \code{\link{DNAbin2genind}} to convert single DNAbin objects.
}
}

