% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.gaps.R
\docType{methods}
\name{rm.gaps}
\alias{rm.gaps}
\alias{rm.gaps,multidna-method}
\alias{rm.gaps,multiphyDat-method}
\title{Remove gap-only sequences for missing data}
\usage{
rm.gaps(x, ...)

\S4method{rm.gaps}{multidna}(x, ...)

\S4method{rm.gaps}{multiphyDat}(x, ...)
}
\arguments{
\item{x}{a \linkS4class{multidna} or \linkS4class{multiphyDat} object.}

\item{...}{further arguments passed to other methods (currently not used).}
}
\description{
In \linkS4class{multidna} and \linkS4class{multiphyDat}, some individuals may not be sequenced for all genes, resulting in gap-only sequences for missing data.
The generic function \code{rm.gaps} has method for both objects; it identifies the missing sequences, and removes gap-only sequences from the alignments wherever needed.
}

