% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,multiphyDat-method}
\alias{show,multiphyDat-method}
\alias{show.multiphyDat}
\title{Display multiphyDat objects}
\usage{
\S4method{show}{multiphyDat}(object)
}
\arguments{
\item{object}{a multiphyDat object}
}
\value{
\code{show} returns an invisible NULL, called for side effects.
}
\description{
Default printing for multiphyDat objects
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}
