% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.gaps.R
\docType{methods}
\name{add.gaps}
\alias{add.gaps}
\alias{add.gaps.generic}
\alias{add.gaps.multidna}
\alias{add.gaps.multiphyDat}
\alias{add.gaps,multidna-method}
\alias{add.gaps,multiphyDat-method}
\title{Add gap-only sequences for missing data}
\usage{
add.gaps(x, ...)

\S4method{add.gaps}{multidna}(x, ...)

\S4method{add.gaps}{multiphyDat}(x, ...)
}
\arguments{
\item{x}{a \linkS4class{multidna} or \linkS4class{multiphyDat} object.}

\item{...}{further arguments passed to other methods (currently not used).}
}
\description{
In \linkS4class{multidna} and \linkS4class{multiphyDat}, some individuals may not be sequenced for all genes.
The generic function \code{add.gaps} has method for both objects; it identifies the missing sequences, and adds gap-only sequences to the alignments wherever needed.
}
