% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doc.R
\docType{package}
\name{apex}
\alias{apex}
\alias{apex-package}
\title{apex: Extensions of Ape for Multiple Gene Data}
\description{
Toolkit for the analysis of multiple gene data. Apex implements the new S4 classes multidna, multiphyDat and associated methods to handle aligned DNA sequences from multiple genes.
}

