% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.multidna.R
\name{read.multidna}
\alias{read.multiFATSA}
\alias{read.multidna}
\title{Read multiple DNA alignments}
\usage{
read.multidna(files, ...)

read.multiFATSA(files)
}
\arguments{
\item{files}{a vector of characters indicating the paths to the files to read from}

\item{...}{further arguments passed to \code{\link[ape]{read.dna}}}
}
\description{
These functions read multiple DNA alignments and store the output in a \linkS4class{multidna} object.
They are relying on ape's original functions \code{\link[ape]{read.dna}} and \code{\link[ape]{read.FASTA}}.


}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}
\seealso{
\code{\link[ape]{read.dna}}, \code{\link[ape]{read.FASTA}}
}

