% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,multidna-method}
\alias{show,multidna-method}
\alias{show.multidna}
\title{Display multidna objects}
\usage{
\S4method{show}{multidna}(object)
}
\arguments{
\item{object}{a multidna object}
}
\description{
Default printing for multidna objects
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}
