% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,multiphyDat-method}
\alias{show,multiphyDat-method}
\title{Display multiphyDat objects}
\usage{
\S4method{show}{multiphyDat}(object)
}
\arguments{
\item{object}{a multiphyDat object}
}
\description{
Default printing for multiphyDat objects
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

